now the footer is a component instead of been part of the appwraper
